+++ CHANGELOG.md
@@ -1,13 +1,23 @@
 
 ## Changelog
 
-#### [Upcoming](https///github.com/kuvaus/changelog-releasenotes-action/compare/v1...HEAD)
+#### [Upcoming](https///github.com/kuvaus/moveclock/compare/v3...HEAD)
 
+#### [v3](https://github.com/kuvaus/moveclock/releases/tag/v3)
 
-#### [v1](https://github.com/kuvaus/changelog-releasenotes-action/releases/tag/v1)
+> 7 May 2024
+
+- Added toggle to move clock before or after System Menu
+
+#### [v2](https://github.com/kuvaus/moveclock/releases/tag/v2)
 
 > 5 May 2024
 
-- statusMenu is now correctly referring to quickSettings (Thanks JustPerfection reviewers)
+- System Menu is now correctly referring to quickSettings (Thanks JustPerfection)
+
+#### [v1](https://github.com/kuvaus/moveclock/releases/tag/v1)
+
+> 5 May 2024
+
 - Initial release